chore: remove filter logic of components & tooltip & crosshair
