+++ docs/dev-demos/src/api/brush.ts
@@ -33,7 +33,6 @@ export const runner = (view: IView) => {
     .legend(view.rootMark)
     .id('rightLegend')
     .scale(colorScale)
-    // .target(markData, 'type')
     .encode({ x: 340, y: 120 });
   view.interaction('legend-filter', {
     source: rightLegend,
