+++ docs/dev-demos/src/api/line.ts
@@ -34,7 +34,6 @@ export const runner = (view: IView) => {
     .legend(view.rootMark)
     .id('rightLegend')
     .scale(colorScale)
-    .target(markData, 'type')
     .encode({ x: 340, y: 160 });
   const xAxis = view
     .axis(view.rootMark)
