+++ docs/dev-demos/src/api/scrollbar-filter.ts
@@ -82,7 +82,6 @@ export const runner = (view: IView) => {
     // .position('top')
     .configure({ 'zIndex': 10 })
     .scale(scrollbarXScale)
-    // .target(markData, 'category')
     .encode({
       range: [0, 0.5],
       x: 40,
