+++ docs/dev-demos/src/api/slider.ts
@@ -47,7 +47,6 @@ export const runner = (view: IView) => {
     .legend(view.rootMark)
     .id('rightLegend')
     .scale(colorScale)
-    // .target(markData, 'type')
     .encode({ x: 340, y: 120 });
   view.interaction('legend-filter', {
     source: rightLegend,
@@ -65,7 +64,6 @@ export const runner = (view: IView) => {
       { scale: datazoomXScale, field: 'category' },
       () => 40
     )
-    // .target(markData, 'category')
     .encode({
       x: 40,
       y: 340,
