+++ docs/dev-demos/src/api/title.ts
@@ -39,13 +39,6 @@ export const runner = (view: IView) => {
       y: 0
     });
 
-  const rightLegend = view
-    .legend(view.rootMark)
-    .id('rightLegend')
-    .scale(colorScale)
-    .target(markData, 'type')
-    .encode({ x: 340, y: 160 });
-
   const xAxis = view
     .axis(view.rootMark)
     .id('xAxis')
@@ -154,54 +147,6 @@ export const runner = (view: IView) => {
     .encode({
       text: (datum: any) => `${datum.amount}`
     });
-  const targetTooltip = view.tooltip(container)
-    .id('targetTooltip')
-    .target(symbol)
-    .title({ value: 'Sales Statistics On Category' })
-    .encode({ offsetX: 10, offsetY: 10 })
-    .content([
-      {
-        key: 'category',
-        value: { field: 'category' },
-        symbol: { fill: 'lightGreen' }
-      },
-      {
-        key: { text: 'amount' },
-        value: datum => datum.amount,
-        symbol: { fill: 'lightGreen', symbolType: 'square' }
-      }
-    ]);
-  const groupTooltip = view.tooltip(container)
-    .id('groupTooltip')
-    .target(line)
-    .title('Total Sales Statistics')
-    .encode({ offsetX: 10, offsetY: 10 })
-    .content([
-      {
-        key: { text: 'amount' },
-        value: datum => datum.amount,
-        symbol: { fill: 'lightGreen', symbolType: 'square' }
-      }
-    ]);
-  const dimensionTooltip = view.dimensionTooltip(container)
-    .id('dimensionTooltip')
-    .scale(xScale)
-    .target(data, 'category')
-    .avoidMark([symbol, line])
-    .title('Sales Statistics On Category')
-    .encode({ offsetX: 10, offsetY: 10 })
-    .content([
-      {
-        key: 'category',
-        value: { field: 'category' },
-        symbol: { fill: 'lightGreen' }
-      },
-      {
-        key: { text: 'amount' },
-        value: datum => datum.amount,
-        symbol: { fill: 'lightGreen', symbolType: 'square' }
-      }
-    ]);
 };
 
 export const callback = (view: IView) => {
