+++ packages/vgrammar-core/src/types/component.ts
@@ -1,12 +1,6 @@
-import type {
-  IGraphicAttribute,
-  ISymbolGraphicAttribute,
-  ITextAttribute,
-  ITextGraphicAttribute
-} from '@visactor/vrender';
+import type { IGraphicAttribute, ITextAttribute } from '@visactor/vrender';
 import type {
   AxisBaseAttributes,
-  BaseCrosshairAttrs,
   BaseLabelAttrs,
   DataLabelAttrs,
   DataZoomAttributes,
@@ -17,9 +11,7 @@ import type {
   PlayerAttributes,
   ScrollBarAttributes,
   SliderAttributes,
-  TitleAttrs,
-  TooltipAttributes,
-  TooltipRowAttrs
+  TitleAttrs
 } from '@visactor/vrender-components';
 import type { AxisEnum, ComponentEnum } from '../graph';
 import type { Nil } from './base';
@@ -27,13 +19,10 @@ import type { IComponent, IData, IGroupMark, IMark, IScale } from './grammar';
 import type {
   ChannelEncodeType,
   ComponentSpec,
-  FieldEncodeType,
   GenerateBasicEncoderSpec,
   MarkFunctionType,
   ScaleEncodeType
 } from './mark';
-import type { IPointLike } from '@visactor/vutils';
-import type { IElement } from './element';
 import type { IView } from './view';
 import type { IBaseScale } from '@visactor/vscale';
 
@@ -103,7 +92,6 @@ export type LegendType = 'auto' | 'discrete' | 'color' | 'size';
 
 export interface ILegend extends IScaleComponent {
   legendType: (legendType: LegendType | Nil) => this;
-  target: (data: IData | string | Nil, filter: string | ((datum: any, legendValues: any) => boolean) | Nil) => this;
 
   // immediate functions
   setSelected: (selectedValues: any[]) => this;
@@ -116,28 +104,6 @@ export interface LegendSpec extends ScaleComponentSpec<LegendBaseAttributes> {
   componentType: ComponentEnum.legend;
   legendType?: LegendType;
   shapeScale?: IScale | string;
-  target?: {
-    data: IData | string;
-    filter: string | ((datum: any, legendValues: any[]) => boolean);
-  };
-}
-
-// crosshair component
-
-export type CrosshairType = 'x' | 'y' | 'angle' | 'radius' | 'radius-polygon';
-
-export type CrosshairShape = 'line' | 'rect';
-
-export interface ICrosshair extends IScaleComponent {
-  crosshairType: (crosshairType: CrosshairType | Nil) => this;
-  crosshairShape: (crosshairShape: CrosshairShape | Nil) => this;
-}
-
-export interface CrosshairSpec extends ScaleComponentSpec<BaseCrosshairAttrs> {
-  componentType: ComponentEnum.crosshair;
-  crosshairType?: CrosshairType;
-  crosshairShape?: CrosshairShape;
-  componentConfig?: { radius?: number; center?: IPointLike };
 }
 
 // slider component
@@ -147,10 +113,6 @@ export type SliderFilterValue = { start: number; end: number };
 export interface ISlider extends IComponent {
   min: (min: MarkFunctionType<number> | Nil) => this;
   max: (max: MarkFunctionType<number> | Nil) => this;
-  target: (
-    data: IData | string | Nil,
-    filter: string | ((datum: any, value: SliderFilterValue) => boolean) | Nil
-  ) => this;
 
   // immediate functions
   setStartEndValue: (start?: number, end?: number) => this;
@@ -160,10 +122,6 @@ export interface SliderSpec extends ComponentSpec<Partial<SliderAttributes>> {
   componentType: ComponentEnum.slider;
   min?: MarkFunctionType<number>;
   max?: MarkFunctionType<number>;
-  target?: {
-    data: IData | string;
-    filter: string | ((datum: any, value: SliderFilterValue) => boolean);
-  };
 }
 
 // datazoom component
@@ -177,10 +135,6 @@ export interface IDatazoom extends IComponent {
     y: ScaleEncodeType | Nil,
     x1?: ChannelEncodeType | Nil,
     y1?: ChannelEncodeType | Nil
-  ) => this;
-  target: (
-    data: IData | string | Nil,
-    filter: string | ((datum: any, value: DatazoomFilterValue) => boolean) | Nil
   ) => this;
 
   // immediate functions
@@ -200,10 +154,6 @@ export interface DatazoomSpec extends ComponentSpec<DataZoomEncoderSpec> {
     y?: ScaleEncodeType;
     x1?: ChannelEncodeType;
     y1?: ChannelEncodeType;
-  };
-  target?: {
-    data: IData | string;
-    filter: string | ((datum: any, value: DatazoomFilterValue) => boolean);
   };
 }
 
@@ -232,7 +182,6 @@ export type PlayerFilterValue = { index: number; value: any };
 export interface IPlayer extends IComponent {
   playerType: (playerType: PlayerType) => this;
   source: (source: IData | string | any[] | Nil) => this;
-  target: (data: IData | string | Nil, source: IData | string | any[] | Nil) => this;
 
   // immediate functions
   play: () => this;
@@ -245,65 +194,6 @@ export interface PlayerSpec extends ComponentSpec<Partial<PlayerAttributes>> {
   componentType: ComponentEnum.player;
   playerType?: PlayerType;
   source?: IData | string | any[];
-  target?: {
-    data: IData | string;
-    source: IData | string | any[];
-  };
-}
-
-// tooltip component
-
-export type CustomTooltipCallback = (
-  datum: any,
-  element: IElement,
-  parameters: any
-) => TooltipRowAttrs | TooltipRowAttrs[];
-
-export interface ITooltipRow {
-  visible?: boolean;
-  key?: MarkFunctionType<string | Partial<ITextGraphicAttribute>> | FieldEncodeType;
-  value?: MarkFunctionType<string | Partial<ITextGraphicAttribute>> | FieldEncodeType;
-  symbol?: MarkFunctionType<string | Partial<ISymbolGraphicAttribute>> | FieldEncodeType;
-}
-
-export interface IBaseTooltip extends IComponent {
-  title: (title: ITooltipRow | string | CustomTooltipCallback | Nil) => this;
-  content: (content: ITooltipRow | ITooltipRow[] | CustomTooltipCallback | Nil) => this;
-}
-
-export interface ITooltip extends IBaseTooltip {
-  target: (mark: IMark | IMark[] | string | string[] | Nil) => this;
-}
-
-export interface BaseTooltipSpec extends ComponentSpec<TooltipAttributes> {
-  title?: ITooltipRow | string | CustomTooltipCallback;
-  content?: ITooltipRow | ITooltipRow[] | CustomTooltipCallback;
-}
-
-export interface TooltipSpec extends BaseTooltipSpec {
-  componentType: ComponentEnum.tooltip;
-  target?: IMark | IMark[] | string | string[];
-}
-
-export type TooltipType = 'x' | 'y' | 'angle' | 'radius';
-
-export interface IDimensionTooltip extends IBaseTooltip {
-  scale: (scale?: IScale | string | Nil) => this;
-  tooltipType: (tooltipType: TooltipType | Nil) => this;
-  target: (data: IData | string | Nil, filter: string | ((datum: any, tooltipValue: any) => boolean) | Nil) => this;
-  avoidMark: (mark: IMark | IMark[] | string | string[] | Nil) => this;
-}
-
-export interface DimensionTooltipSpec extends BaseTooltipSpec {
-  componentType: ComponentEnum.dimensionTooltip;
-  scale?: IScale | string;
-  tooltipType?: TooltipType;
-  target?: {
-    data: IData | string;
-    filter: string | ((datum: any, tooltipValue: any) => boolean);
-  };
-  avoidMark?: IMark | IMark[] | string | string[];
-  componentConfig?: { center?: IPointLike };
 }
 
 // title component
@@ -324,10 +214,6 @@ export interface TitleSpec extends ComponentSpec<Partial<TitleAttrs>> {
 export type ScrollbarFilterValue = { start?: number; end?: number; startRatio: number; endRatio: number };
 
 export interface IScrollbar extends IScaleComponent {
-  target: (
-    data: IData | string | Nil,
-    filter: string | ((datum: any, value: ScrollbarFilterValue) => boolean) | Nil
-  ) => this;
   container: (container: IGroupMark | string | Nil) => this;
   direction: (direction: MarkFunctionType<Direction> | Nil) => this;
   position: (position: MarkFunctionType<OrientType> | Nil) => this;
@@ -338,10 +224,6 @@ export interface IScrollbar extends IScaleComponent {
 
 export interface ScrollbarSpec extends ScaleComponentSpec<Partial<ScrollBarAttributes>> {
   componentType: ComponentEnum.scrollbar;
-  target?: {
-    data: IData | string;
-    filter: string | ((datum: any, value: ScrollbarFilterValue) => boolean);
-  };
   container?: IGroupMark | string;
   direction?: MarkFunctionType<Direction>;
   position?: MarkFunctionType<OrientType>;
@@ -352,13 +234,10 @@ export interface ScrollbarSpec extends ScaleComponentSpec<Partial<ScrollBarAttri
 export type BuiltInComponentSpec =
   | AxisSpec
   | LegendSpec
-  | CrosshairSpec
   | SliderSpec
   | DatazoomSpec
   | LabelSpec
   | PlayerSpec
-  | TooltipSpec
-  | DimensionTooltipSpec
   | TitleSpec
   | ScrollbarSpec;
