+++ packages/vgrammar-full/src/index.ts
@@ -17,12 +17,10 @@ import {
   // components
   registerAxis,
   registerLegend,
-  registerCrosshair,
   registerSlider,
   registerLabel,
   registerDataZoom,
   registerPlayer,
-  registerTooltip,
   registerTitle,
   registerGrid,
   registerScrollbar,
@@ -85,12 +83,10 @@ import {
 View.useRegisters([
   registerAxis,
   registerLegend,
-  registerCrosshair,
   registerSlider,
   registerLabel,
   registerDataZoom,
   registerPlayer,
-  registerTooltip,
   registerTitle,
   registerGrid,
   registerScrollbar,
