+++ packages/vgrammar-plot/__tests__/util.ts
@@ -3,12 +3,10 @@ import {
   // components
   registerAxis,
   registerLegend,
-  registerCrosshair,
   registerSlider,
   registerLabel,
   registerDataZoom,
   registerPlayer,
-  registerTooltip,
   registerTitle,
   registerGrid,
   registerScrollbar,
@@ -71,12 +69,10 @@ import {
 View.useRegisters([
   registerAxis,
   registerLegend,
-  registerCrosshair,
   registerSlider,
   registerLabel,
   registerDataZoom,
   registerPlayer,
-  registerTooltip,
   registerTitle,
   registerGrid,
   registerScrollbar,
