+++ packages/vgrammar/src/index.ts
@@ -3,12 +3,10 @@ import {
   // components
   registerAxis,
   registerLegend,
-  registerCrosshair,
   registerSlider,
   registerLabel,
   registerDataZoom,
   registerPlayer,
-  registerTooltip,
   registerTitle,
   registerGrid,
   registerScrollbar,
@@ -95,12 +93,10 @@ View.useRegisters([
   // components
   registerAxis,
   registerLegend,
-  registerCrosshair,
   registerSlider,
   registerLabel,
   registerDataZoom,
   registerPlayer,
-  registerTooltip,
   registerTitle,
   registerGrid,
   registerScrollbar,
